refactor(matrices): move conversion methods to RepMatrix
